Core/Misc: Reduce code differences between branches
(cherry picked from commit 69b647a24ca5c86772409550bb9df8e9f79531ec)

+++ src/server/game/Spells/SpellInfo.cpp
@@ -465,7 +465,7 @@ bool SpellEffectInfo::IsAura() const
 
 bool SpellEffectInfo::IsAura(AuraType aura) const
 {
-    return IsAura() && ApplyAuraName == uint32(aura);
+    return IsAura() && ApplyAuraName == aura;
 }
 
 bool SpellEffectInfo::IsTargetingArea() const
